SR-IOV for mlx5 interfaces: control and model changes
Changes in control/interfaces.py and model/interfaces.py to
support the new API called 'action' and a new key to be
retrieved by the lookup called 'actions'.

At this moment there's only one action available in the
backend: action 'SR-IOV' of interfaces with the module
mlx5_core/mlx5-core.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>