Move System Modules to a sidebar entry
The System Services tab is not only for system services. It also handles
the System Modules features.

As those feature are not related, there is no reason to keep them
together.

This patch adds a new sidebar entry for System Modules and split the
code into 2 HTML files.

This patch does not change any UI element. It is just moving code around
to get the features separated into 2 different sidebar entries.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>